Operational F4 pipeline & price pipeline dev mode
Spent hours debugging feature pipeline for form 4, mainly the issues
were datatypes discrpeancies between feature store and the kafka topic
as well as not allowing spark time to process the data before the next,
leading to concurrency limit issues.

On the other hand, the price pipeline has also been debugged and is now
able to fetch data from our feature store and request to yahoo only the
data missing. Furthermore, the requests are made on batches of tickers
which are got from the feature store (form 4 database), avoiding this way
unnecessary requests to yahoo (only for the tickers we have transactions
from.)

Next steps will be dockerizing the price pipeline and ensuring everything
works fine (testing).