Create 3rd price-pipeline microservice for avg.
Developed the third price-pipeline microservice intended to compute the
average of the percentage change of the stock price for each ticker.

For each ticker, working with a df of transactions and updating it
with a new column ('avg_target') that contains the average of the change
in price. It considers only those transactions whose date + delta are
before today's date.

Have also standarized some env & config names across microservices.