Target microservice operational :)
Delta Offset tracker implemented
Connection handles feature group initializations
Minor coherence fixes across committed files

Delta offset enables track guarantee that the system does not process
the same data twice. This is achieved by storing the last processed
timestamp in the feature store and using it to filter out data that.
To avoid repeated API calls to hopsworks, the offsets are sent in buffer
to the feature store and are only written to the store when the buffer
is full.