Microservice integration into a single system
Refactored microservices of the sec-feature-pipeline and prices-feature-
pipeline into docker compose files. Now, the system can run in inference
mode and training mode without adjusting anything.

If willing to modify the behavior, the environment variables associated
to each system type can be modified.

Developed a naming convention for all the feature stores and topics.
This will allow for the system to be more scalable and easier to manage.

Erased duplicated code such as the feature store and removed unused modules.

+ Lots of debugging and testing